feat: unwrap command namespace to flat exports + barrel
